refactor(number-theory): isolate friable contracts
