+++ src/jacobian/math/number_theory/__init__.py
@@ -1,7 +1,7 @@
 """Supported exact number-theory API."""
 
+from jacobian.math.number_theory._friable_models import FriableCountResult
 from jacobian.math.number_theory._friable_operations import count_friable
-from jacobian.math.number_theory._models import FriableCountResult
 from jacobian.math.number_theory.ramanujan_sums import ramanujan_sum
 
 __all__ = ["FriableCountResult", "count_friable", "ramanujan_sum"]
